requirement.txt
python-dotenv
pandas 
numpy
pymongo
certifi 
pymongo[srv]==3.12



